Move map and sidebar related components into sub folders since we will add more and more components now
