+++ src/sidebar/Search.module.css
