+++ src/sidebar/Sidebar.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from 'react'
 import { ApiInfo, Instruction, Path } from '@/routing/Api'
 import { RouteStoreState } from '@/stores/RouteStore'
 import { QueryStoreState } from '@/stores/QueryStore'
-import Search from '@/Search'
-import styles from '@/Sidebar.module.css'
+import Search from '@/sidebar/Search'
+import styles from '@/sidebar/Sidebar.module.css'
 import Dispatcher from '@/stores/Dispatcher'
 import { SetSelectedPath } from '@/actions/Actions'
